+++ src/routes/(v2)/v2/Navigation.svelte
@@ -27,6 +27,7 @@
 		'submit-json',
 		'trim-fields',
 		'unions',
+		'valibot-dates',
 		'validate',
 		'validate-update',
 		'validators-clear'